Added, readjusted, and removed content on about page
Continued to make progress on About me page, removing content styling and elements that didn't align with the final vision I had for the site. Added embeded video, styled it to match the rest of the page. Added media query for mobile and tablet sized screens that changes the page layout. Made color adjustments to the header and card elements.