Making adjustments to header nav icons
Switched out the img tags for the svg tag itself, to allow for more flexibility for customization. Experimenting with different ideas for how I want to display to users what page you're on using the nav icon elements, still havent decided which one I want for it. stay tuned for more